Migrated patient health serializer to new file and added new fields
